NO-TICKET: execution-engine: move runtime_context into a dir
